refactor(ContinuousMovementRule): 패키지 분리
