+++ README.md
@@ -3,6 +3,7 @@
 ## 페어와 지킬 컨벤션
 1. 클래스 정의 다음 줄은 공백으로 한다.
 2. test code에 사용하는 메서드는 `static import`한다.
+3. `this`는 같은 클래스의 객체가 파라미터로 넘어왔을 때, 파라미터 변수 명이 필드의 변수 명과 겹칠 때 사용한다.
 
 ## 기능 요구 사항
