+++ src/main/java/chess/domain/movement/continuous/EastMovement.java
@@ -1,16 +1,16 @@
-package chess.domain.movement;
+package chess.domain.movement.continuous;
 
 import chess.domain.position.Position;
 
-public class NorthMovement extends ContinuousMovementRule {
+public class EastMovement extends ContinuousMovementRule {
 
     @Override
     protected boolean isMovable(int rankDifference, int fileDifference) {
-        return fileDifference == 0 && rankDifference > 0;
+        return rankDifference == 0 && fileDifference > 0;
     }
 
     @Override
     protected Position next(Position position) {
-        return position.moveToNorth();
+        return position.moveToEast();
     }
 }
