+++ src/main/java/chess/domain/movement/continuous/NorthMovement.java
@@ -1,16 +1,16 @@
-package chess.domain.movement;
+package chess.domain.movement.continuous;
 
 import chess.domain.position.Position;
 
-public class EastMovement extends ContinuousMovementRule {
+public class NorthMovement extends ContinuousMovementRule {
 
     @Override
     protected boolean isMovable(int rankDifference, int fileDifference) {
-        return rankDifference == 0 && fileDifference > 0;
+        return fileDifference == 0 && rankDifference > 0;
     }
 
     @Override
     protected Position next(Position position) {
-        return position.moveToEast();
+        return position.moveToNorth();
     }
 }
