+++ src/test/java/chess/domain/movement/continuous/NorthMovementTest.java
@@ -1,4 +1,4 @@
-package chess.domain.movement;
+package chess.domain.movement.continuous;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,36 +8,36 @@ import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class EastMovementTest {
+class NorthMovementTest {
 
     @Test
     @DisplayName("이동 가능한지 확인한다.")
     void isMovableTest() {
         Position start = new Position(File.C, Rank.FOUR);
-        Position end = new Position(File.E, Rank.FOUR);
-        EastMovement eastMovement = new EastMovement();
+        Position end = new Position(File.C, Rank.SIX);
+        NorthMovement northMovement = new NorthMovement();
 
-        assertThat(eastMovement.isMovable(start, end)).isTrue();
+        assertThat(northMovement.isMovable(start, end)).isTrue();
     }
 
     @Test
     @DisplayName("이동 불가능한지 확인한다.")
     void isMovableTest_false() {
         Position start = new Position(File.C, Rank.FOUR);
-        Position end = new Position(File.E, Rank.THREE);
-        EastMovement eastMovement = new EastMovement();
+        Position end = new Position(File.B, Rank.SIX);
+        NorthMovement northMovement = new NorthMovement();
 
-        assertThat(eastMovement.isMovable(start, end)).isFalse();
+        assertThat(northMovement.isMovable(start, end)).isFalse();
     }
 
     @Test
     @DisplayName("이동 경로를 알 수 있다.")
     void findPathTest() {
         Position start = new Position(File.C, Rank.FOUR);
-        Position end = new Position(File.E, Rank.FOUR);
-        EastMovement eastMovement = new EastMovement();
+        Position end = new Position(File.C, Rank.SIX);
+        NorthMovement northMovement = new NorthMovement();
 
-        assertThat(eastMovement.findPath(start, end))
-                .containsExactly(new Position(File.D, Rank.FOUR), new Position(File.E, Rank.FOUR));
+        assertThat(northMovement.findPath(start, end))
+                .containsExactly(new Position(File.C, Rank.FIVE), new Position(File.C, Rank.SIX));
     }
 }
